added jeremie's astri/chec converter functions
